refactor(capabilities): thin the tcp cluster root to the engine cluster shape
Mirror the aether.component cluster's decomposition (component/route.rs
precedent) so tcp/mod.rs is a thin declaration file like engine/mod.rs:

- Move the send-side facades (TcpWasmExt, TcpNativeExt, the peer
  resolvers, and the private session_mailbox_id helper) into a new
  tcp/route.rs; the root re-exports both traits so their public paths
  are unchanged.
- Move the inline test module into a sibling tcp/tests.rs, declared
  cfg(all(test, feature = "runtime")).
- Narrow tcp/config.rs's gate from not(target_family = "wasm") to
  feature = "runtime": the configs are child-actor init bundles holding
  raw TcpListener / TcpStream handles, consumed only by the
  runtime-gated halves; comments state the rationale.
- Re-point the stale ext-helper comment onto the always-on
  HandlesKind<MonitorNotice> marker that actually justifies the import.

Closes #2477
